+++ R/ggeffect.R
@@ -4,7 +4,7 @@
 #' @description
 #'   \code{ggeffect()} computes marginal effects of model terms. It internally
 #'   calls \code{\link[effects]{Effect}} and puts the result into tidy data
-#'   frames. \code{eff()} is an alias for \code{ggeffect()}.
+#'   frames.
 #'
 #' @param model A fitted model object, or a list of model objects. Any model
 #'   that is supported by the \CRANpkg{effects}-package should work.
@@ -221,11 +221,4 @@ ggeffect_helper <- function(model, terms, ci.lvl, x.as.factor, ...) {
   if (!x.as.factor) mydf$x <- sjlabelled::as_numeric(mydf$x, keep.labels = FALSE)
 
   mydf
-}
-
-
-#' @rdname ggeffect
-#' @export
-eff <- function(model, terms, ci.lvl = .95, x.as.factor = FALSE, ...) {
-  ggeffect(model, terms, ci.lvl, x.as.factor, ...)
 }