+++ NEWS.md
@@ -2,11 +2,13 @@
 
 ## General
 
-* Removed alias names `mem()` and `ame()`.
+* Removed alias names `mem()`, `eff()` and `ame()`.
+* For mixed models (packages **lme4**, **nlme**), the uncertainty of the random effect variances is now taken into account when `type = "re"`.
 
 ## Changes to functions
 
 * `ggpredict()` gets a `condition`-argument to specify values at which covariates should be held constant, instead of their `typical` value.
+* The `terms`-argument in `ggpredict()` can now select a range of feasible values for numeric values, e.g. `terms = "age[range]"`.
 * `plot()` gets some new arguments to control which plot-title to show or hide: `show.title`, `show.x.title` and `show.y.title`.
 * `plot()` gets a `log.y` argument to transform the y-axis to logarithmic scale, which might be useful for binomial models with predicted probabilities.
 * The `plot()`-method for plotting all effects with `ggpredict()` (when `term = NULL`) now allows to arrange the plot in facets (using `facets = TRUE`).
@@ -15,6 +17,7 @@
 ## Bug fixes
 
 * Fixed issue with binomial models that used `cbind(...)` as response variable.
+* Fixed issue with suboptimal precision of confidence resp. prediction intervals for mixed models (packages **lme4**, **nlme**), which are now more accurate.
 
 # ggeffects 0.3.4